requirements.txt
Flask==2.1.3
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
torch==1.12.0
torchvision==0.13.0

telegram==0.0.1